+++ plantcv/geospatial/analyze/dsm.py
@@ -1,5 +1,5 @@
 # Analyze Digital Surface Model (DSM) over many regions
-from plantcv.geospatial._helpers import _gather_ids
+from plantcv.geospatial._helpers import _gather_ids, _set_nodata_term
 from plantcv.plantcv import outputs, params
 from rasterio.plot import plotting_extent
 from matplotlib import pyplot as plt
@@ -29,10 +29,7 @@ def height_percentile(dsm, geojson, lower=25, upper=90, label=None):
     # Cast to float since zonal_stats gives overflow error on uint8 data
     dsm_data = dsm_data.astype(np.float32)
 
-    if dsm.metadata['nodata'] is not None:
-        nodata_value = dsm.metadata['nodata']
-    else:
-        nodata_value = -999
+    nodata_value = _set_nodata_term(dsm)
     # Scale of the data
     scale = dsm.metadata["crs"].linear_units
 
@@ -46,17 +43,15 @@ def height_percentile(dsm, geojson, lower=25, upper=90, label=None):
     region_upper_avgs = zonal_stats(geojson, dsm_data,
                                     affine=dsm.metadata["transform"],
                                     nodata=nodata_value, stats=upper)
-    # Gather plot IDs from the geojson
-    ids = _gather_ids(geojson=geojson)
-
-    # Set lable to params.sample_label if None
+    # Set lable to params.sample_label if no other labels provided
     if label is None:
-        label = params.sample_label
+        # Gather plot IDs from the geojson
+        label = _gather_ids(geojson=geojson)
 
     soil_vals = []
     plant_vals = []
     # Save data to outputs
-    for i, id_lbl in enumerate(ids):
+    for i, id_lbl in enumerate(label):
         # Initialize no data cases
         avg1, avg2, avg = [0.0, 0.0, nodata_value]
         # Save soil heights
@@ -66,7 +61,7 @@ def height_percentile(dsm, geojson, lower=25, upper=90, label=None):
                                 trait="dsm_mean_below_" + str(lower),
                                 method="plantcv-geospatial.analyze.dsm",
                                 scale=scale, datatype=float,
-                                value=avg1, label=label)
+                                value=avg1, label=scale)
         soil_vals.append(avg1)
         # Save plant heights
         if region_upper_avgs[i][upper] is not None:
@@ -75,7 +70,7 @@ def height_percentile(dsm, geojson, lower=25, upper=90, label=None):
                                 trait="dsm_mean_above_" + str(upper),
                                 method="plantcv-geospatial.analyze.dsm",
                                 scale=scale, datatype=float,
-                                value=avg2, label=label)
+                                value=avg2, label=scale)
         plant_vals.append(avg2)
         if avg1 != 0 and avg2 != 0:
             avg = avg2 - avg1
@@ -83,7 +78,7 @@ def height_percentile(dsm, geojson, lower=25, upper=90, label=None):
                                 trait="height",
                                 method="plantcv-geospatial.analyze.dsm",
                                 scale=scale, datatype=float,
-                                value=avg, label=label)
+                                value=avg, label=scale)
 
     # Min and max height of plots
     min_elevation = min(soil_vals)