+++ docs/changelog.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented below.
 
 #### geospatial.analyze.coverage
 
-* v0.1dev: **geospatial.analyze.coverage**(*img, bin_mask, geojson*)
+* v0.1dev: **geospatial.analyze.coverage**(*img, bin_mask, geojson, label=None*)
 
 #### geospatial.analyze.height_percentile
 
 * v0.1dev: **geospatial.analyze.height_percentile**(*dsm, geojson, lower=25, upper=90, label=None*)
+
+#### geospatial.analyze.spectral_index
+
+* v0.1dev: **geospatial.analyze.spectral_index**(*img, geojson, label=None*)
 
 #### geospatial.points2roi_circle